Add more comments and improve code
